Renamed restService.c to serverStatusService to clarify ambiguity of function
